modified:   controllers/kategoricontroller/kategoricontroller.go 	modified:   controllers/produkcontroller/produkcontroller.go 	renamed:    logic/logic.go -> util/util.go
