+++ util/util.go
@@ -1,8 +1,9 @@
-package logic
+package util
 
 import (
-	"html/template"
 	"net/http"
+	"strconv"
+	"text/template"
 )
 
 func RenderTemplate(w http.ResponseWriter, tmpl string, data interface{}) {
@@ -11,4 +12,12 @@ func RenderTemplate(w http.ResponseWriter, tmpl string, data interface{}) {
 		panic(err)
 	}
 	temp.Execute(w, data)
+}
+
+func ParsingId(idString string) (int, error) {
+	id, err := strconv.Atoi(idString)
+	if err != nil {
+		return 0, err
+	}
+	return id, nil
 }
